+++ src/utils/iconInit.ts
@@ -3,6 +3,8 @@ import { faCalendarAlt } from "@fortawesome/free-solid-svg-icons/faCalendarAlt";
 import { faCat } from "@fortawesome/free-solid-svg-icons/faCat";
 import { faCheck } from "@fortawesome/free-solid-svg-icons/faCheck";
 import { faChevronDown } from "@fortawesome/free-solid-svg-icons/faChevronDown";
+import { faCircleCheck } from "@fortawesome/free-solid-svg-icons/faCircleCheck";
+import { faCircleInfo } from "@fortawesome/free-solid-svg-icons/faCircleInfo";
 import { faClock } from "@fortawesome/free-solid-svg-icons/faClock";
 import { faDog } from "@fortawesome/free-solid-svg-icons/faDog";
 import { faFileAudio } from "@fortawesome/free-solid-svg-icons/faFileAudio";
@@ -12,6 +14,7 @@ import { faInfoCircle } from "@fortawesome/free-solid-svg-icons/faInfoCircle";
 import { faSpinner } from "@fortawesome/free-solid-svg-icons/faSpinner";
 import { faTimes } from "@fortawesome/free-solid-svg-icons/faTimes";
 import { faTrash } from "@fortawesome/free-solid-svg-icons/faTrash";
+import { faTriangleExclamation } from "@fortawesome/free-solid-svg-icons/faTriangleExclamation";
 
 export const initializeIconList = () => {
   const iconList = [
@@ -28,6 +31,10 @@ export const initializeIconList = () => {
     faTimes,
     faTrash,
     faClock,
+
+    faTriangleExclamation,
+    faCircleCheck,
+    faCircleInfo,
   ] as const;
 
   // Set up icon libraries